Cache MAC address to speed up uuid1/0
Closes #40.

Getting the MAC address (through :inet.getifaddrs/0) is slow, resulting
in the uuid1/0 function being around 100x slower than the other UUID
functions. With this commit we're starting up a GenServer that caches
the MAC address in :persistent_term if available or ETS otherwise.
This GenServer refreshes the MAC address every 10 seconds.